Move existing SF tests to sqlparser_snowflake.rs
