+++ tests/sqlparser_snowflake.rs
@@ -9,9 +9,14 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+#[macro_use]
+#[path = "macros/mod.rs"]
+mod macros;
+
+use sqlparser::test_utils::*;
 use sqlparser::ast::*;
 use sqlparser::dialect::{GenericDialect, SnowflakeDialect};
-use sqlparser::test_utils::*;
 use sqlparser::tokenizer::*;
 
 #[test]
@@ -61,6 +66,45 @@ fn test_snowflake_single_line_tokenize() {
     ];
 
     assert_eq!(expected, tokens);
+}
+
+#[test]
+fn test_sf_derives_single_table_in_parenthesis() {
+    // Nesting a subquery in parentheses is non-standard, but supported in Snowflake SQL
+    let sql = "SELECT * FROM ((SELECT 1) AS t)";
+    let select = snowflake_and_generic().verified_only_select(sql);
+    let from = only(select.from);
+    assert_eq!(
+        from.relation,
+        TableFactor::NestedJoin(Box::new(TableWithJoins {
+            relation: TableFactor::Derived {
+                lateral: false,
+                subquery: Box::new(snowflake_and_generic().verified_query("SELECT 1")),
+                alias: Some(TableAlias {
+                    name: "t".into(),
+                    columns: vec![],
+                })
+            },
+            joins: Vec::new(),
+        }))
+    );
+}
+
+#[test]
+fn test_single_table_in_parenthesis() {
+    // Parenthesized table names are non-standard, but supported in Snowflake SQL
+    let sql = "SELECT * FROM (a NATURAL JOIN (b))";
+    let select = snowflake_and_generic().verified_only_select(sql);
+    let from = only(select.from);
+
+    assert_eq!(from.relation, nest!(table("a"), nest!(table("b"))));
+
+    // Double parentheses around table names are non-standard, but supported in Snowflake SQL
+    let sql = "SELECT * FROM (a NATURAL JOIN ((b)))";
+    let select = snowflake_and_generic().verified_only_select(sql);
+    let from = only(select.from);
+
+    assert_eq!(from.relation, nest!(table("a"), nest!(nest!(table("b")))));
 }
 
 fn snowflake_and_generic() -> TestedDialects {
